added a style class
